positionFeeMapper 코드명과 함께 출력

+++ src/main/java/com/dartmedia/dmss/mapper/PositionFeeMapperMapper.java
@@ -6,5 +6,5 @@ import com.dartmedia.dmss.dto.PositionFeeMapper;
 
 @MapperInterface
 public interface PositionFeeMapperMapper extends IMapper<PositionFeeMapper> {
-  public PositionFeeMapper readByCode(int positionCode, int feeCode) throws Exception;
+  public PositionFeeMapper readByCode(short positionCode, short feeCode) throws Exception;
 }
